requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
arrow==1.2.3
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.0
blessed==1.20.0
certifi @ file:///work/ci_py311/certifi_1676822443922/work/certifi
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.0
colorama==0.4.6
croniter==1.3.8
dateutils==0.6.12
deepdiff==6.3.0
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.88.0
filelock==3.10.0
frozenlist==1.3.3
fsspec==2023.3.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.13.2
idna==3.4
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lightning==2.0.0
lightning-cloud==0.5.31
lightning-utilities==0.8.0
lit==15.0.7
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
ordered-set==4.1.0
orjson==3.8.7
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.1.1
pretty-errors==1.2.25
psutil==5.9.4
pydantic==1.10.6
Pygments==2.14.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.0
pytz==2022.7.1
PyYAML==6.0
readchar==4.0.5
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
rich==13.3.2
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.95
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.22.0
starsessions==1.3.0
sympy==1.11.1
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==2.0.0
torchaudio==2.0.1
torchmetrics==0.11.4
torchvision==0.15.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.1
triton==2.0.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
wcwidth==0.2.6
websocket-client==1.5.1
websockets==10.4
yarl==1.8.2
